Android: Cast JNI pointer to jint instead of jlong
A VideoCaptureDeviceAndroid pointer was being cast to jlong instead of
jint. The usual way is to cast to jint, which has sufficient size for
pointers.

BUG=246979

Review URL: https://chromiumcodereview.appspot.com/15793008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@204365 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 51d4724ffd049b6b373f7bef74ed6e50781304ff

+++ media/video/capture/android/video_capture_device_android.cc
@@ -106,7 +106,7 @@ bool VideoCaptureDeviceAndroid::Init() {
 
   j_capture_.Reset(Java_VideoCapture_createVideoCapture(
       env, base::android::GetApplicationContext(), id,
-      reinterpret_cast<jlong>(this)));
+      reinterpret_cast<jint>(this)));
 
   return true;
 }
